+++ src/admin.ts
@@ -2,6 +2,7 @@
 import crossFetch from 'cross-fetch';
 import * as Types from './types';
 import { coerceInt64Fields, hasWindow, trimURL } from './utils';
+import { toSDKError } from './errors';
 
 // set fetch based on window object. Cross fetch have issues with umd build
 const getFetcher = () => (hasWindow() ? window.fetch : crossFetch);
@@ -33,21 +34,7 @@ const scimEndpointFragment = 'id org_id enabled created_at updated_at';
 
 function toErrorList(errors: unknown): Types.AuthorizerSDKError[] {
   if (Array.isArray(errors)) {
-    return errors.map((item) => {
-      if (item instanceof Error) return item;
-      if (item && typeof item === 'object' && 'message' in item) {
-        const err: Types.AuthorizerSDKError = new Error(
-          String((item as { message: unknown }).message),
-        );
-        const extensions = (item as { extensions?: unknown }).extensions;
-        if (extensions && typeof extensions === 'object') {
-          const code = (extensions as { code?: unknown }).code;
-          if (typeof code === 'string') err.code = code;
-        }
-        return err;
-      }
-      return new Error(String(item));
-    });
+    return errors.map(toSDKError);
   }
   if (errors instanceof Error) return [errors];
   if (errors !== null && typeof errors === 'object') {
